Letterforms and Skeleton Updates
I worked on updating the letterforms and took a more geometric direction, my thinking is that since we are applying a roughness filter, cleaner forms will do the job and make it easier to complete the typeface without having to deal with spacing and shape issues from the handwritten style.

I redesigned every letter in the string and made sure to solve the issue of similar letters not looking related by making thigs as simple as I can, I introduced crossbars that shoot out on the other side of the stem and much curvier curves on the t and f, so it bears some semblance to Nsibidi. I believe from this point on, layering Nsibidi radicals on these glyphs would be much straight forward. I believe I still have a lot of work to do on spacing however. I would fix that in my next release.