Began Implementing bracket layers
I finally figured out how to impliment bracket layers for all my three masters and implemented it on the letters A and B, so far nothing broke and the glyphs are compatible.

Next step is figuring out reverse bracket layers and then implementing that across all my basic glyphs so I can export and test a stable version of the variable font.